Establish and enforce an integer-conversion convention (#133) (#157)
* Add integer cast convention design spec (#133)


* Address spec review findings in cast-convention design (#133)


* Add cast-convention implementation plan; ground spec details (#133)

Spec amendments from plan grounding: picture_type flips to u32 (FLAC's
picture type field is natively u32; only mp3 APIC narrows, deliberately),
and the fuse mtime carve-out is dropped (clippy does not flag the guarded
cast, so an #[expect] would be unfulfilled).


* Address plan review findings in cast-convention plan (#133)


* Add sanctioned u64->usize conversion behind a 64-bit guard (#133)


* Make Track audio bounds unsigned; validate at row-read (#133)


* Make art/binary-tag byte quantities unsigned (#133)


* Make ordinals u64 and picture_type u32 across db rows (#133)


* Clear remaining cast-lint sites in musefs-db (#133)


* Apply cast convention across musefs-core (#133)


* Route db row I/O through rusqlite's checked conversions (#133)

rusqlite's FromSql/ToSql impls for u64/usize sit behind the fallible_uint
feature; without it the row boundary was hand-rolled (manual < 0 checks,
bare sign-changing casts) and width/height/picture_type/ordinal reads
silently wrapped negative values instead of erroring. Enable the feature
and let typed gets/params do the validation, per the cast-convention spec.


* Apply cast convention to FLAC/WAV/VorbisComment synthesis (#133)


* Apply cast convention to MP3/ID3 synthesis (#133)


* Make SYNCHSAFE_MAX u32 to drop filter-site casts (#133)


* Apply cast convention to MP4 synthesis (#133)


* Collapse the redundant new_moov_size guard into its try_from (#133)

The guard and the size-field try_from enforced the same u32 bound, leaving
the try_from error arm unreachable (an unkillable mutant for the in-diff
mutation gate). The try_from is now the single check.


* Apply cast convention to Ogg synthesis and fixtures (#133)


* Apply cast convention to fuse adapter and latencyfs (#133)


* Use expect for the guard-bounded fuse mtime conversion (#133)


* Apply cast convention to test and bench fixtures (#133)


* Enforce the integer cast convention via clippy lints (#133)


* Kill or document surviving mutants from the cast-convention diff (#133)

Refresh mutants.toml anchors the migration's line drift broke, document
the new synchsafe_decode |^= equivalence class, and add tests for the
real gaps (track_art_images, scanned art dimensions, ogg seq_delta,
build_packets value-len boundary, flac_block 24-bit length). Exclude two
newly-surfaced equivalents (wav walk_chunks advance guard, fixtures::wav
RIFF-size sum) with documented reasons.


* Exclude the machine-killing usize_from constant-return mutants (#133)

Replacing musefs-format's u64->usize helper body with a constant pins
parser walk positions while their collect-loops push per iteration: an
allocation bomb that OOMs the host (26.7G RSS + 11.5G swap observed)
before cargo-mutants' test timeout fires — it killed two local tmux
sessions and two CI runners. Exclude that mutant pair with the analysis
documented; harden CLAUDE.md's local-gate instructions (disk-backed
TMPDIR + memory-capped cgroup; /tmp is RAM-backed tmpfs here).


* Pin the vorbis_body_empty fixture contract with a format-local test (#133)

Its consumers are musefs-core tests, but the mutation gate runs each
crate's own suite, so only a format-local test can kill mutations of the
fixture body.


---------

+++ musefs-core/tests/probe_equivalence.rs
@@ -11,10 +11,10 @@ use musefs_db::Db;
 /// `(sha256, picture_type, description, ordinal)`.
 type NormalizedTrack = (
     String,
-    i64,
-    i64,
-    Vec<(String, String, i64)>,
-    Vec<(String, i64, String, i64)>,
+    u64,
+    u64,
+    Vec<(String, String, u64)>,
+    Vec<(String, u32, String, u64)>,
 );
 
 /// Normalize a DB to comparable rows: tracks by path, tags by (key,value,ordinal),
